Make hitboxes arbitrary polygons

+++ core/src/com/po/fuck/Entity.java
@@ -8,13 +8,13 @@ public class Entity extends GameObject {
     protected Movement movement = null;
     protected Weapon weapon = null;
 
-    protected int health_points;
-    public final int MAX_HEALTH_POINTS;
+    protected float health_points;
+    public final float MAX_HEALTH_POINTS;
     protected boolean immortal = false;
 
     public HealthBar healthBar;
 
-    public Entity(Vector2 position,int HP) {
+    public Entity(Vector2 position, float HP) {
         super(position);
         this.MAX_HEALTH_POINTS = HP;
         this.health_points = HP;
@@ -25,7 +25,7 @@ public class Entity extends GameObject {
         return immortal || health_points > 0;
     }
 
-    public boolean takeDamage(int damage) {
+    public boolean takeDamage(float damage) {
         if (immortal)
             return false;
